requirements.txt
torch
torchvision
torchmetrics
Pillow
scikit-learn
scikit-image
transformers
tqdm
tabulate
timm
matplotlib
opencv-python